* move remove role from user to roles controller

+++ app/controllers/roles_controller.rb
@@ -32,11 +32,30 @@ class RolesController < ApplicationController
   end
   
   def destroy
-    r = Role.find_all_by_id(params[:id])
-    if r && !r.empty?
-      r.first.destroy
+    if params[:id] && params[:user_id] 
+      remove_role_from_user params[:id], params[:user_id]
+    else
+      r = Role.find_all_by_id(params[:id])
+      if r && !r.empty?
+        r.first.destroy
+      end
     end
     redirect_to roles_path
+  end
+
+  private
+  def remove_role_from_user role_id, user_id
+    # remove role from user
+    begin
+      r = Role.find(role_id)
+      u = User.find(user_id)
+      if u && r
+        u.roles_users.select{|ru| ru.role_id == r.id}.map(&:destroy)
+        flash[:notice] = "Removed #{r.role} role from #{u.fullname}"
+      end
+    rescue ActiveRecord::RecordNotFound => ex
+      flash[:error] = "Unable to find the role or user.  Nothing done."
+    end
   end
 
 end
